support patching of pg-query-stream object
pg-query-stream is a Class, so it can't be patched as seamlessly as the client.
Instead, we apply the patch to a new method called 'named', which is also a Class.  This allows the users to call it the same way as before using 'new'.  The patch is also returned, so it could actually be used directly.  Here are few examples:

-- #1 Patch During Require --
  var PgQueryStream = require('pg-query-stream')
  var QueryStream  = require('node-postgres-named').patch(PgQueryStream)
-- or --
  var QueryStream  = require('node-postgres-named').patch(require('pg-query-stream'))

  let query = new QueryStream('SELECT * FROM generate_series(1, $qty) num', {qty: 200})
  let stream = client.query(query)

-- #2 use the patched 'named' method --
  var PgQueryStream = require('pg-query-stream')
  var named = require('node-postgres-named')

  named.patch(PgQueryStream)
  let query = new PgQueryStream.named('SELECT * FROM generate_series(1, $qty) num', {qty: 200})
  let stream = client.query(query)